[flutter_tools] make archive extraction failure a tool exit
